requirements.txt
langchain>=0.0.267
openai>=0.27.8
numpy>=1.24.3
matplotlib>=3.7.1
python-dotenv>=1.0.0